requirements.txt
django>=4.2,<4.3
djangorestframework>=3.14.0,<3.15.0
psycopg2-binary==2.9.6
# psycopg2>=2.7.5, <2.8.0